Implement authentication user in album CRUD

+++ app/Http/Controllers/V1/AlbumController.php
@@ -16,40 +16,48 @@ class AlbumController extends Controller
      */
     public function index(Request $request)
     {
-        return Album::all();
+        return Album::where('user_id', $request->user()->id)->get();
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \App\Http\Requests\AlbumRequest  $request
+     * @param \App\Http\Requests\AlbumRequest $request
      * @return \Illuminate\Http\Response
      */
     public function store(AlbumRequest $request)
     {
-        return Album::create($request->all());
+        $data = $request->all();
+        $data['user_id'] = $request->user()->id;
+        return Album::create($data);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Album  $album
+     * @param \App\Models\Album $album
      * @return \Illuminate\Http\Response
      */
-    public function show(Album $album)
+    public function show(Request $request, Album $album)
     {
+        if ($album->user_id !== $request->user()->id) {
+            return abort(403, 'Unauthorized action.');
+        }
         return $album;
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \App\Http\Requests\AlbumRequest  $request
-     * @param  \App\Models\Album  $album
+     * @param \App\Http\Requests\AlbumRequest $request
+     * @param \App\Models\Album $album
      * @return \Illuminate\Http\Response
      */
     public function update(AlbumRequest $request, Album $album)
     {
+        if ($album->user_id !== $request->user()->id) {
+            return abort(403, 'Unauthorized action.');
+        }
         $album->update($request->all());
         return $album;
     }
@@ -57,11 +65,14 @@ class AlbumController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Album  $album
+     * @param \App\Models\Album $album
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Album $album)
+    public function destroy(Request $request, Album $album)
     {
+        if ($album->user_id !== $request->user()->id) {
+            return abort(403, 'Unauthorized action.');
+        }
         $album->delete();
     }
 }
